[BI-1720] - merged redundant BrAPIStudyDAO
