+++ src/main/java/org/breedinginsight/brapi/v2/services/BrAPIStudyService.java
@@ -21,7 +21,7 @@ import io.micronaut.http.server.exceptions.InternalServerException;
 import lombok.extern.slf4j.Slf4j;
 import org.brapi.client.v2.model.exceptions.ApiException;
 import org.brapi.v2.model.core.BrAPIStudy;
-import org.breedinginsight.brapi.v2.dao.BrAPIStudyDAO;
+import org.breedinginsight.brapps.importer.daos.BrAPIStudyDAO;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
